feat(TextField): separated the component logic and interface into separate files
