+++ src/components/inputs/TextField/stories/TextField.Valid.stories.js
@@ -1,6 +1,6 @@
 import { TextField } from "..";
 import { TextFieldController } from "./TexfieldController";
-import { sizes } from "..";
+import { sizes } from "../interface";
 import {
   parameters,
   label,
